Don't crash when secondary dependencies are checked
Now that a class is used instead of a string for projects, __eq__
previously only worked for other projects. But checking secondary
dependencies compared a string which crashed.

Notes: This doesn't handle transitive dependencies of internal projects
that shadow a public package correctly. It still checks the public
packages requirements.